backend/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
billiard==4.1.0
blinker==1.6.2
cachelib==0.9.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
command-not-found==0.3
contourpy==1.1.0
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
dnspython==2.4.1
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Caching==2.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
fonttools==4.42.0
greenlet==2.0.2
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
importlib-resources==6.0.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
kiwisolver==1.4.4
kombu==5.3.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.25.2
oauthlib==3.2.0
packaging==23.1
passlib==1.7.4
Pillow==10.0.0
prompt-toolkit==3.0.39
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==5.4.1
redis==4.6.0
redis2==2.10.6.3
requests==2.31.0
SecretStorage==3.3.1
six==1.16.0
sql-alchemy-db==0.0.4
SQLAlchemy==2.0.19
systemd-python==234
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.4
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.6
Werkzeug==2.3.6
WTForms==3.0.1
zipp==1.0.0
